Use correct target in refund transaction (#4)
[Here](https://cowservices.slack.com/archives/C0371SB243E/p1705896112132899)
we saw an eth-flow transaction that couldn't be refunded by this script
because it was created in barn and not in prod (they use different
contract addresses).

This code changes the `to` address of the transaction to be the one
derived from the logs.

I also noticed that the script only works for the simple case where an
EOA creates an order by directly calling the eth-flow contract and would
fail in more complex transactions (for example, an eth-flow order
created by a Safe). This is not caused by the changes in this PR,
however they make it easy to detect that this is happening.
I added a simple check for that so that an error message makes this
issue immediately evident.