Support non-direct interactions (#7)
Currently, the script doesn't support cases when a user transferred ETH
using an intermediate SC.
This PR addresses it by searching for a proper item in the tx receipt
using the OrderPlacement event hash.

### Changes

- In case the recipient address is not equals to the ETHflow SC, that
means an intermediate interaction was involved in the tx.
- In that case, the tx data can't be used directly since it contains
additional interactions.
- Search for the `createOrder` function selector, extract and parse the
`EthFlowOrder`, which the `invalidateOrder` function expected.
- Encode the `invalidateOrder` function.
- The remaining process is the same.

### How to test


https://github.com/cowprotocol/manual-ethflow-refunder/pull/7#issuecomment-2747700682